928 pivot changes
